feat: added components and data to access collections
